+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Changing an email address to 'Primary' in account form now changes the rest to 'Other'
 * Changed account API so it can deal with related fields
 * Placeholders for email are simplified, `http://` is no longer required
+* When replying in the historylist, if the email is flagged inactive under account/contact, it tries to find another emailaddress to reply to
+* 
 
 #### Bug Fixes